+++ src/main/java/kitchenpos/menu/domain/JdbcTemplateMenuGroupDao.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.menu.domain;
 
-import kitchenpos.model.MenuGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
