+++ src/main/java/kitchenpos/menu/domain/MenuBo.java
@@ -1,12 +1,7 @@
-package kitchenpos.bo;
+package kitchenpos.menu.domain;
 
-import kitchenpos.dao.MenuDao;
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.dao.MenuProductDao;
-import kitchenpos.dao.ProductDao;
-import kitchenpos.model.Menu;
-import kitchenpos.model.MenuProduct;
-import kitchenpos.model.Product;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductDao;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
