+++ src/main/java/kitchenpos/menu/domain/MenuGroupDao.java
@@ -1,6 +1,4 @@
-package kitchenpos.dao;
-
-import kitchenpos.model.MenuGroup;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
