+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.model;
+package kitchenpos.menu.domain;
 
 public class MenuProduct {
     private Long seq;
